+++ tests/test_create_config/test_modules/test_meta_config.py
@@ -1,22 +1,22 @@
 from unittest.mock import patch
 
-from pyxspress.create_config.modules.create_meta_liveView import (
+from pyxspress.create_config.modules.meta_config import (
     _live_view,
     _meta_writer_endpoints,
     _meta_writer_sensor,
-    live_view_file,
-    meta_writer_file,
+    create_live_view_launch_script,
+    create_meta_writer_launch_script,
 )
 
 
-def test_meta_writer_sensor():
+def test_meta_writer_sensor() -> None:
     num_chans = 4
     sensor = _meta_writer_sensor(num_chans)
     expected = f"--sensor-shape {num_chans} 4096"
     assert sensor == expected
 
 
-def test_meta_writer_endpoints():
+def test_meta_writer_endpoints() -> None:
     num_cards = 4
     endpoints = _meta_writer_endpoints(num_cards)
     for card in range(num_cards):
@@ -25,7 +25,7 @@ def test_meta_writer_endpoints():
         assert ep_str in endpoints
 
 
-def test_live_view():
+def test_live_view() -> None:
     num_cards = 4
     live_view = _live_view(num_cards)
     for card in range(num_cards):
@@ -34,12 +34,12 @@ def test_live_view():
         assert live_view_str in live_view
 
 
-@patch("pyxspress.create_config.modules.create_meta_liveView._live_view")
+@patch("pyxspress.create_config.modules.meta_config._live_view")
 @patch("os.chmod")
-def test_live_view_file(mock_chmod, mock_live_view, template_dir, tmp_path):
+def test_live_view_file(mock_chmod, mock_live_view, template_dir, tmp_path) -> None:
     num_cards = 2
     mock_live_view.return_value = "15500,15501"
-    live_view_file(num_cards, template_dir, tmp_path)
+    create_live_view_launch_script(num_cards, template_dir, tmp_path)
     assert mock_live_view.called
     assert (tmp_path / "stLiveViewMerge.sh").exists()
     text = (tmp_path / "stLiveViewMerge.sh").read_text()
@@ -48,8 +48,8 @@ def test_live_view_file(mock_chmod, mock_live_view, template_dir, tmp_path):
     assert mock_chmod.called
 
 
-@patch("pyxspress.create_config.modules.create_meta_liveView._meta_writer_endpoints")
-@patch("pyxspress.create_config.modules.create_meta_liveView._meta_writer_sensor")
+@patch("pyxspress.create_config.modules.meta_config._meta_writer_endpoints")
+@patch("pyxspress.create_config.modules.meta_config._meta_writer_sensor")
 @patch("os.chmod")
 def test_meta_writer_file(
     mock_chmod,
@@ -57,14 +57,14 @@ def test_meta_writer_file(
     mock_meta_writer_endpoints,
     template_dir,
     tmp_path,
-):
+) -> None:
     num_cards = 2
     num_chans = 4
 
     mock_meta_writer_sensor.return_value = "sensor response"
     mock_meta_writer_endpoints.return_value = "endpoints response"
 
-    meta_writer_file(num_cards, num_chans, template_dir, tmp_path)
+    create_meta_writer_launch_script(num_cards, num_chans, template_dir, tmp_path)
     assert mock_meta_writer_sensor.called
     assert mock_meta_writer_endpoints.called
     assert (tmp_path / "stMetaWriter.sh").exists()